\name{gen_univar_NOTAMO}
\alias{gen_univar_NOTAMO}

\title{
Generate univariate Notamo sample
}
\description{
Generates a sample from a distribution with prespecified moments. Requires the output of \code{\link{estimate_quantile_mix}}.
}
\usage{
gen_univar_NOTAMO(quantile_mix_out,N)
}

\arguments{
  \item{quantile_mix_out}{
output of \code{\link{estimate_quantile_mix}}
  }
  \item{N}{
sample size
  }
}
\details{
}
\value{
}
\references{
}
\author{
Max Auerswald and Morten Moshagen
}
\note{
}


\seealso{
\code{\link{estimate_quantile_mix}}, \code{\link{gen_multivar_NOTAMO}}
}
\examples{

}

